pyproject.toml
[project]
name = "bp-ecg-preprocessing-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dask>=2025.1.0",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "pymupdf>=1.25.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "surya-ocr>=0.8.3",
]
